alert

+++ app/src/main/java/iti/android/wheatherappjetpackcompose/dataLayer/source/dto/Alert.kt
@@ -4,10 +4,10 @@ import com.google.gson.annotations.SerializedName
 
 data class Alert(
     val description: String?,
-    val end: Int?,
+    val end: Long?,
     val event: String?,
     @SerializedName("sender_name")
     val senderName: String?,
-    val start: Int?,
+    val start: Long?,
     val tags: List<String> = emptyList(),
 )
